add proxy for amazonaws
Issue with proxying to aws. Prior to download there is a server
certificate handshake that proxying does not allow. Next plan is to try
this out on my home server to add private keys to the proxy. These will
need to be added via environment variables

It should look like the following on a curl:

```
* TCP_NODELAY set
* Connected to github-production-release-asset-2e65be.s3.amazonaws.com
* (54.231.72.243) port 443 (#0)
* TLS 1.2 connection using TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256
* Server certificate: *.s3.amazonaws.com
* Server certificate: DigiCert Baltimore CA-2 G2
* Server certificate: Baltimore CyberTrust Root
```

This does not appear to be available during requests made via proxy from
the curl that I am outputting. This is my current guess on the
webdriver-manager not being able to pull down gecko on an update command
with a proxy.